Added vanity gen, fixed some missing functions
